refactor: add place entity

+++ app/src/main/java/com/untilled/roadcapture/data/datasource/dao/PictureDao.kt
@@ -27,5 +27,5 @@ interface PictureDao {
     fun initThumbnail() : Completable
 
     @Query("SELECT IFNULL(MAX(`order`), 0) FROM picture")
-    fun getNextOrder(): Single<Long>
+    fun getNextOrder(): Single<Int>
 }
