+++ app/src/main/java/com/untilled/roadcapture/data/repository/picture/PictureRemoteRepository.kt
@@ -18,5 +18,5 @@ interface PictureRemoteRepository {
 
     fun initThumbnail() : Completable
 
-    fun getNextOrder() : Single<Long>
+    fun getNextOrder() : Single<Int>
 }
